+++ src/testing-utils/if-type/booleans/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @file Entry point for boolean IfTypes.
+ */
+
+export * from './if-boolean';
+export * from './if-boolean-broad';
+export * from './if-boolean-literal';
+export * from './if-true';
+export * from './if-false';