+++ src/testing-utils/if-type/index.ts
@@ -5,3 +5,4 @@
 export * from './if';
 
 export * from './any-never-unknown';
+export * from './booleans';